fixed so you can hold a card and click again over dropzone it detects the dropzone
commented out some dropzone text replacing it with a white zone during testing, where you can drop the cards to dropzone.  TODO need to make a script to every card that hold the data for suit and rank. So when you drop the card in the zone it knows which caard it is