Done with start changed with prefab and sprites etc. They work different now look in discripttion
card now starts with a cardblueprint prefab. Then got 2 child gameobject with the visuals and 2 gameobject with number. Changed so now depending on if they have face up or down the sprite can change etc. Only seen that it works on start will work on it more tomorrow. I am TIRED. 2025-02-19 21:25